+++ cimd/cli.py
@@ -4,7 +4,7 @@ import click
 
 from cimd.classes import context
 from cimd.commands import add, delete, get, list
-from cimd.commands.scratch import pipeline_logs
+from cimd.commands.extract import pipeline_logs
 
 
 @click.group(context_settings={"auto_envvar_prefix": "CIMD", "help_option_names": ["-h", "--help"]})
@@ -26,15 +26,15 @@ def cli(ctx: click.core.Context, file: str, debug: bool) -> None:
     ctx.obj = context.ApplicationContext(filename=file, debug=debug)
 
 
-@click.group(name="scratch")
-def scratch_group() -> click.Group:  # type: ignore[empty-body]
-    """Collect and add metadata from sources."""
+@click.group(name="extract")
+def extract_group() -> click.Group:  # type: ignore[empty-body]
+    """Scape and collect metadata from sources."""
 
 
-scratch_group.add_command(pipeline_logs.pipeline_logs_command)
+extract_group.add_command(pipeline_logs.pipeline_logs_command)
 
 cli.add_command(add.add_command)
 cli.add_command(delete.delete_command)
 cli.add_command(get.get_command)
 cli.add_command(list.list_command)
-cli.add_command(scratch_group)
+cli.add_command(extract_group)